Step 2 - Move really JMS specific tests to new PushingJmsListenerTest
